rename testfaucet to faucet
