+++ packages/integration-tests/test/addTokenTest.js
@@ -16,7 +16,7 @@ const RootChain = require('@omisego/omg-js-rootchain')
 const ChildChain = require('@omisego/omg-js-childchain')
 const Web3 = require('web3')
 
-const faucet = require('../helpers/testFaucet')
+const faucet = require('../helpers/faucet')
 const config = require('../test-config')
 const rcHelper = require('../helpers/rootChainHelper')
